requirements.txt
Flask==3.0.3
Werkzeug==3.0.3
python-dotenv==1.0.1
pydantic==2.7.4
google-generativeai==0.7.2
reportlab==4.2.2
gunicorn==23.0.0
PyPDF2==3.0.1
